chage codition

+++ c/factorial.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 // fuction declearation for the factorail function
-int factorial(int n);
+long int factorial(int n);
 
 // void main function
 
@@ -12,11 +12,11 @@ void main()
     printf("enter nunmber for the factorial");
     scanf("%d", &n);
     int i = factorial(n);
-    printf("the inputed number is %d\n", i);
+    printf("the inputed number is %ld\n", i);
 }
 
 // factorail function for full calculation
-int factorial(int n)
+long int factorial(int n)
 {
     if (n >= 1)
     {
